+++ spec/dummy/app/models/article.rb
@@ -3,6 +3,6 @@
 class Article < ActiveRecord::Base
   has_logidze
 
-  belongs_to :user, Rails::VERSION::MAJOR >= 5 ? {optional: true} : {}
+  belongs_to :user, optional: true
   has_many :comments
 end